+++ src/components/GenreList.tsx
@@ -1,9 +1,21 @@
-import { HStack, Image, List, ListItem, Spinner, Text } from '@chakra-ui/react';
-import useGenres from '../hooks/useGenres';
+import {
+  Button,
+  HStack,
+  Image,
+  List,
+  ListItem,
+  Spinner,
+} from '@chakra-ui/react';
+import useGenres, { Genre } from '../hooks/useGenres';
 import getCroppedImageUrl from '../services/image-url';
-import { MdEnergySavingsLeaf } from 'react-icons/md';
 
-const GenreList = () => {
+// #region Interface
+interface IGenreList {
+  onSelectedGenre: (genre: Genre) => void;
+}
+// #endregion
+
+const GenreList = ({ onSelectedGenre }: IGenreList) => {
   const { data: genres, error, isLoading } = useGenres();
 
   if (error) return null;
@@ -20,7 +32,13 @@ const GenreList = () => {
               borderRadius={8}
               src={getCroppedImageUrl(genre.image_background)}
             />
-            <Text fontSize="lg">{genre.name}</Text>
+            <Button
+              fontSize="lg"
+              variant="link"
+              onClick={() => onSelectedGenre(genre)}
+            >
+              {genre.name}
+            </Button>
           </HStack>
         </ListItem>
       ))}