+++ src/hooks/useData.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import apiClient from "../services/api-client";
-import { CanceledError } from "axios";
+import { AxiosRequestConfig, CanceledError } from "axios";
 // <T> is a generic type that will be inferred by TypeScript
 
 
@@ -13,7 +13,7 @@ interface FetchResponse<T> {
 }
 // #endregion
 
-const useData = <T>(endpoint:string) => {
+const useData = <T>(endpoint:string, requestConfig?: AxiosRequestConfig, deps?: any) => {
   const [data, setData] = useState<T[]>([]);
   const [error, setError] = useState<string>('');
   const [isLoading, setLoading] = useState<boolean>(false);
@@ -22,7 +22,7 @@ const useData = <T>(endpoint:string) => {
     const controller = new AbortController();
     setLoading(true);
     apiClient
-      .get<FetchResponse<T>>(endpoint, { signal: controller.signal })
+      .get<FetchResponse<T>>(endpoint, { signal: controller.signal, ...requestConfig })
       .then((response) => {
         setData(response.data.results);
         setLoading(false);
@@ -33,7 +33,7 @@ const useData = <T>(endpoint:string) => {
         setLoading(false);
       });
     return () => controller.abort();
-  }, []);
+  }, deps? [...deps]: []);
 
   return { data, error, isLoading };
 };